added some minor features alert window on delete

+++ src/components/todo-app/Todo.jsx
@@ -32,6 +32,9 @@ export default function Todo() {
     }
 
     const handledelete = async (id) => {
+
+        let confirm = (window.confirm("ARE YOU SURE?"))
+        if (confirm == false) return
         await axios.delete(`http://localhost:8080/api/${id}`).then((response) => {
             setreaload(!reload)
         }).catch((error) => {
